--- memes.js
@@ -1,1 +1,0 @@
-// This file should contain functions for scraping memes off various websites.
